Duplicate `Open/CreateFILE` to each .cpp files
* Sort headers
* Add `[[noreturn]]` attribute to `OnJPEGError`
