+++ Auxiliary/DirectXTexPNG.cpp
@@ -11,16 +11,53 @@
 
 #include "DirectXTexPNG.h"
 
-#include <filesystem>
 #include <cstdio>
+#include <filesystem>
 #include <png.h>
 
 namespace DirectX
 {
     using std::filesystem::path;
+    using ScopedFILE = std::unique_ptr<FILE, int(*)(FILE*)>;
 
-    std::unique_ptr<FILE, int(*)(FILE*)> OpenFILE(const path& p) noexcept(false);
-    std::unique_ptr<FILE, int(*)(FILE*)> CreateFILE(const path& p) noexcept(false);
+    namespace
+    {
+    #if defined(_WIN32)
+        ScopedFILE OpenFILE(const path& p) noexcept(false)
+        {
+            const std::wstring fpath = p.generic_wstring();
+            FILE* fp = nullptr;
+            if (auto ec = _wfopen_s(&fp, fpath.c_str(), L"rb"); ec)
+                throw std::system_error{ ec, std::system_category(), "_wfopen_s" };
+            return { fp, &fclose };
+        }
+        ScopedFILE CreateFILE(const path& p) noexcept(false)
+        {
+            const std::wstring fpath = p.generic_wstring();
+            FILE* fp = nullptr;
+            if (auto ec = _wfopen_s(&fp, fpath.c_str(), L"w+b"); ec)
+                throw std::system_error{ ec, std::system_category(), "_wfopen_s" };
+            return { fp, &fclose };
+        }
+    #else
+        ScopedFILE OpenFILE(const path& p) noexcept(false)
+        {
+            const std::string fpath = p.generic_string();
+            FILE* fp = fopen(fpath.c_str(), "rb");
+            if (fp == nullptr)
+                throw std::system_error{ errno, std::system_category(), "fopen" };
+            return { fp, &fclose };
+        }
+        ScopedFILE CreateFILE(const path& p) noexcept(false)
+        {
+            const std::string fpath = p.generic_string();
+            FILE* fp = fopen(fpath.c_str(), "w+b");
+            if (fp == nullptr)
+                throw std::system_error{ errno, std::system_category(), "fopen" };
+            return { fp, &fclose };
+        }
+    #endif
+    }
 
     [[noreturn]] void OnPNGError(png_structp, png_const_charp msg)
     {
